+++ src/anemoi/registry/commands/upload.py
@@ -1,16 +1,12 @@
-#!/usr/bin/env python
-# (C) Copyright 2024 ECMWF.
+# (C) Copyright 2024 Anemoi contributors.
 #
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
 # In applying this licence, ECMWF does not waive the privileges and immunities
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
-#
 
-"""Command place holder. Delete when we have real commands.
-
-"""
 
 import logging
 import os
@@ -40,7 +36,7 @@ class Upload:
         if not UPLOAD_ALLOWED:
             LOG.error("Direct upload not allowed.")
             return
-        from anemoi.utils.s3 import upload
+        from anemoi.utils.remote.s3 import upload
 
         upload(args.path, args.target, overwrite=args.overwrite)