+++ src/anemoi/registry/entry/dataset.py
@@ -1,9 +1,12 @@
-# (C) Copyright 2023 European Centre for Medium-Range Weather Forecasts.
+# (C) Copyright 2024 Anemoi contributors.
+#
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
 # In applying this licence, ECMWF does not waive the privileges and immunities
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
+
 
 import datetime
 import logging
@@ -12,6 +15,7 @@ import os
 import yaml
 from anemoi.datasets import open_dataset
 from anemoi.utils.humanize import when
+from anemoi.utils.sanitise import sanitise
 
 from anemoi.registry import config
 from anemoi.registry.rest import RestItemList
@@ -36,8 +40,27 @@ class DatasetCatalogueEntry(CatalogueEntry):
     collection = COLLECTION
     main_key = "name"
 
+    @classmethod
+    def publish(cls, path):
+        PLATFORM = config()["datasets_platform"]
+        STATUS = "experimental"
+
+        entry = DatasetCatalogueEntry(path=path)
+        entry.register()
+        entry.set_status(STATUS)
+
+        recipe = entry.record["metadata"].get("recipe", {})
+        if recipe:
+            entry.set_recipe(recipe)
+        else:
+            LOG.warning("No recipe found in metadata.")
+
+        target = entry.build_location_path(PLATFORM)
+        entry.upload(path, target, platform=PLATFORM)
+        entry.add_location(PLATFORM, target)
+
     def set_status(self, status):
-        self.rest_item.patch([{"op": "add", "path": "/status", "value": status}])
+        self.patch([{"op": "add", "path": "/status", "value": status}])
 
     def build_location_path(self, platform, uri_pattern=None):
         if uri_pattern is None:
@@ -49,14 +72,18 @@ class DatasetCatalogueEntry(CatalogueEntry):
         return uri_pattern.format(name=self.key)
 
     def add_location(self, platform, path):
+        if not path.startswith("s3://"):
+            path = os.path.abspath(path)
+            path = os.path.normpath(path)
+
         LOG.debug(f"Adding location to {platform}: {path}")
-        self.rest_item.patch([{"op": "add", "path": f"/locations/{platform}", "value": {"path": path}}])
+        self.patch([{"op": "add", "path": f"/locations/{platform}", "value": {"path": path}}])
         return path
 
     def remove_location(self, platform, *, delete):
         if delete:
             self.delete(platform)
-        self.rest_item.patch([{"op": "remove", "path": f"/locations/{platform}"}])
+        self.patch([{"op": "remove", "path": f"/locations/{platform}"}])
 
     def delete(self, platform):
         if not config().get("allow_delete"):
@@ -67,7 +94,7 @@ class DatasetCatalogueEntry(CatalogueEntry):
             LOG.warning(f"Nothing to delete for {self.key} on platform {platform}")
             return
         if path.startswith("s3://"):
-            from anemoi.utils.s3 import delete
+            from anemoi.utils.remote.s3 import delete
 
             return delete(path + "/")
         else:
@@ -113,7 +140,7 @@ class DatasetCatalogueEntry(CatalogueEntry):
         self.transfer(task, source_path, target, resume=True, threads=2)
 
     def transfer(self, task, source_path, target, resume, threads):
-        from anemoi.utils.s3 import upload
+        from anemoi.utils.remote import transfer
 
         from anemoi.registry.workers.transfer_dataset import Progress
 
@@ -121,30 +148,38 @@ class DatasetCatalogueEntry(CatalogueEntry):
         LOG.info(f"Upload('{source_path}','{target}', resume=True, threads=2)")
         task.set_status("running")
         try:
-            upload(source_path, target, resume=resume, threads=threads, progress=progress)
+            transfer(source_path, target, resume=resume, threads=threads, progress=progress)
         except:
             task.set_status("stopped")
             raise
         task.unregister()
 
-    def set_recipe(self, file):
-        if not os.path.exists(file):
-            raise FileNotFoundError(f"Recipe file not found: {file}")
+    def _file_or_dict(self, file):
+        if isinstance(file, dict):
+            return file
         if not file.endswith(".yaml"):
             LOG.warning("Recipe file extension is not .yaml")
         with open(file) as f:
-            recipe = yaml.safe_load(f)
-        self.rest_item.patch([{"op": "add", "path": "/recipe", "value": recipe}])
+            return yaml.safe_load(f)
+
+    def set_recipe(self, file):
+        recipe = self._file_or_dict(file)
+        self.patch([{"op": "add", "path": "/metadata/recipe", "value": sanitise(recipe)}])
+
+    def set_variables_metadata(self, file):
+        variables_metadata = self._file_or_dict(file)
+        self.patch([{"op": "add", "path": "/metadata/variables_metadata", "value": variables_metadata}])
 
     def load_from_path(self, path):
         import zarr
 
         if not path.startswith("/") and not path.startswith("s3://"):
             LOG.warning(f"Dataset path is not absolute: {path}")
+            path = os.path.abspath(path)
         if not os.path.exists(path) and not path.startswith("s3://"):
-            LOG.warning(f"Dataset path does not exist: {path}")
+            raise ValueError(f"Dataset path does not exist: {path}")
         if not path.endswith(".zarr") or path.endswith(".zip"):
-            LOG.warning("Dataset path extension is neither .zarr nor .zip")
+            raise ValueError(f"Dataset path extension is not supported ({path})")
 
         name, _ = os.path.splitext(os.path.basename(path))