+++ src/anemoi/registry/entry/experiment.py
@@ -1,9 +1,12 @@
-# (C) Copyright 2023 European Centre for Medium-Range Weather Forecasts.
+# (C) Copyright 2024 Anemoi contributors.
+#
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
 # In applying this licence, ECMWF does not waive the privileges and immunities
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
+
 
 import datetime
 import logging
@@ -12,9 +15,9 @@ import tempfile
 from getpass import getuser
 
 import yaml
-from anemoi.utils.s3 import delete
-from anemoi.utils.s3 import download
-from anemoi.utils.s3 import upload
+from anemoi.utils.remote.s3 import delete
+from anemoi.utils.remote.s3 import download
+from anemoi.utils.remote.s3 import upload
 
 from anemoi.registry.rest import RestItemList