+++ CHANGELOG.md
@@ -11,10 +11,22 @@ Keep it human-readable, your future self will thank you!
 ## [Unreleased]
 
 ### Added
+
+- Add anemoi-transform link to documentation
 - CI workflows to check for updates in the changelog and the documentation.
+- Support for "anemoi-datasets publish"
+- Added set from file (python only)
+- Added 'update' command
+- Force full paths when registering
+- Added naming conventions
+- Fix docstring errors
+- Fix import errors [#18](https://github.com/ecmwf/anemoi-registry/pull/18)
+- Remove usage of obsolete upload function from anemoi-utils.
 
 ### Changed
 - Replaces the deploy workflow with cd-pypi
+- Update copyright notice
+- Fix `__version__` import in init
 
 ### Removed